requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==0.4.3
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
braceexpand==0.1.7
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
einops==0.3.0
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.0.0
gradio==3.13.2
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
imhist==0.0.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
imWatermark==0.0.2
invisible-watermark==0.1.5
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
lightning-utilities==0.10.1
linkify-it-py==2.0.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.4
mdit-py-plugins==0.4.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.15.0
onnxruntime==1.16.3
open-clip-torch==2.7.0
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
orjson==3.9.12
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.2
pudb==2019.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
Pympler==1.0.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytorch-lightning==1.4.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.1
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
semver==3.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
-e git+https://github.com/withSang/stable-diffusion-poc-gprc-server@20856c032564920de07de04c1ee7f6611de4822e#egg=stable_diffusion
starlette==0.35.1
streamlit==1.30.0
streamlit-drawable-canvas==0.8.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.6.0
torchvision==0.13.1
tornado==6.4
tqdm==4.66.1
transformers==4.19.2
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2023.4
tzlocal==5.2
uc-micro-py==1.0.2
urllib3 @ file:///croot/urllib3_1698257533958/work
urwid==2.5.0
uvicorn==0.27.0.post1
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
